+++ src/graphics/Font.h
@@ -3,7 +3,6 @@
 
 #include <vector>
 #include <map>
-#include "Point.h"
 #include "PathSegment.h"
 
 class Font {
@@ -11,11 +10,11 @@ public:
 	Font(const char* fontFilePath, double scale);
 
   	std::vector<PathSegment> getCharacterPath(char character);
-	long getCharacterWidth(char character);
+	double getCharacterWidth(char character);
 
   private:
 	std::map<char, std::vector<PathSegment> > characterPaths;
-	std::map<char, long> characterWidths;
+	std::map<char, double> characterWidths;
 };
 
 #endif
